Get persistent data from localStorage and mini refactor
