+++ src/actions/index.js
@@ -1,4 +1,6 @@
-export const expandInfo = payload => ({
-    type: 'FILTER_BY_REGION',
-    payload,
-})
+export const expandInfo = (payload) => ({
+  type: 'FILTER_BY_REGION',
+  payload,
+});
+
+export default {};
