+++ src/utils/saveToLocalStorage.js
@@ -1,0 +1,8 @@
+export default function saveToLocalStorage(state) {
+  try {
+    const serializedState = JSON.stringify(state);
+    localStorage.setItem('state', serializedState);
+  } catch (e) {
+    console.log(e);
+  }
+};
